refactor(extensions): group UI and workspace support
