+++ src/extensions/workspace/workspaceWriteGuard.test.ts
