+++ src/extensions/workspace/workspaceWriteGuard.ts
