+++ src/ui/ext/extensionDialogs.test.ts
