+++ src/ui/ext/extensionDocumentReader.ts
