+++ src/ui/ext/extensionNavigation.test.ts
