+++ src/ui/ext/extensionNotifications.test.ts
