+++ src/ui/ext/extensionPaintTheme.ts
