+++ src/ui/ext/extensionPanes.test.ts
