+++ src/ui/ext/extensionPanes.ts
