+++ src/ui/ext/extensionSelection.test.ts
