+++ src/ui/ext/extensionTrustPrompt.ts
