+++ src/ui/ext/synchronousExtensionCallback.ts
